regen after fix for float types in generator for 5.0 and 5.1 apis

+++ azure-devops/azure/devops/v5_0/contributions/models.py
@@ -689,7 +689,7 @@ class ResolvedDataProvider(Model):
     """ResolvedDataProvider.
 
     :param duration: The total time the data provider took to resolve its data (in milliseconds)
-    :type duration: int
+    :type duration: float
     :param error:
     :type error: str
     :param id:
@@ -697,7 +697,7 @@ class ResolvedDataProvider(Model):
     """
 
     _attribute_map = {
-        'duration': {'key': 'duration', 'type': 'int'},
+        'duration': {'key': 'duration', 'type': 'float'},
         'error': {'key': 'error', 'type': 'str'},
         'id': {'key': 'id', 'type': 'str'}
     }
